web-ecommerce-app/api/urls.py
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from rest_framework.routers import DefaultRouter
from api.register import RegisterViewSet
from api.products import ProductViewSet

router = DefaultRouter()
router.register(r'register', RegisterViewSet, basename='register')
router.register(r'products', ProductViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('browser-auth/', include('rest_framework.urls')),
    path('auth/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
