web-ecommerce-app/requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
mysqlclient==2.0.3
Pillow==8.3.2
PyJWT==2.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
